Move other RSpec configs into separate files.
