+++ spec/factories/user_factory.rb
@@ -5,7 +5,6 @@ FactoryGirl.define do
   factory :user do
     sequence(:email) { |n| "user#{n}@email.com" }
     password "password"
-    password_confirmation "password"
 
     profile { Profile.new username: generate(:username) }
   end
