+++ spec/support/shoulda_matchers.rb
@@ -1,0 +1,13 @@
+require "shoulda-matchers"
+
+
+RSpec.configure do |config|
+
+  Shoulda::Matchers.configure do |config|
+    config.integrate do |with|
+      with.test_framework :rspec
+      with.library :rails
+    end
+  end
+
+end
